xo generates update params as pointers properly + ideas for gen

+++ internal/repos/postgresql/postgresqltestutil/user.go
@@ -36,6 +36,6 @@ func RandomUserCreateParams(t *testing.T) db.UserCreateParams {
 		LastName:   pointers.New(testutil.RandomLastName()),
 		ExternalID: testutil.RandomString(10),
 		Scopes:     []string{"scope1", "scope2"},
-		RoleRank:   int16(testutil.RandomInt(2, 4)),
+		RoleRank:   int16(testutil.RandomInt64(2, 4)),
 	}
 }
